requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
PyGithub==1.59.1
anthropic==0.7.8
python-multipart==0.0.6
httpx==0.27.0
pydantic==1.10.13

# Optional but highly recommended for testing
pytest==8.1.1
respx==0.20.0
pytest-asyncio==0.23.6
